Update README.md
I created this project using ARKit and RealityKit (primarily), which is able to automatically generate thumbnails for usdz files and place selected models into the ARView, with guidance from the central square (from the Focus Entity) indicating where to place the model. The placed model can be manipulated through various operations such as drag and zoom. It is a basic use case of ARKit and RealityKit with relevant resources, and will be extended to include more features in the future.